requirements.txt
concurrent-log-handler==0.9.22
Django==4.0.2
django-environ==0.4.5
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
factory_boy
gunicorn==20.1.0
Pillow==10.0.0
pytz==2022.6
psycopg2-binary==2.8.6
pytest==6.2.5
pytest-django==4.5.2
requests==2.27.1
requests-toolbelt==0.10.1